backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.1"
dependencies = [
    "django==6.0.0",
    "djangorestframework==3.16.1",
    "psycopg==3.1.18",
    "python-dotenv==1.0.1",
    "djangorestframework-simplejwt==5.5.0",
    "django-cors-headers==4.4.0",
    "celery==5.4.0",
    "redis==5.0.7",
    "setuptools==80.9.0",
    "django-unfold>=0.72.0",
    "django-import-export>=4.3.14",
]
